refactor: divide some login into hooks
