+++ src/components/CharacterList/CharacterList.tsx
@@ -1,30 +1,25 @@
 /* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
 import { CharacterItem } from '@components/CharacterItem/CharacterItem';
-import { Character } from '@models/index';
+import { useHandleDetails } from '@hooks/useHandleDetails';
+import { CharacterWithId } from '@models/index';
 import classnames from 'classnames';
-import { useLocation, useNavigate } from 'react-router-dom';
 import styles from './CharacterList.module.scss';
 
 interface CharacterListProps {
-  characters: Character[];
+  characters: CharacterWithId[];
   isDetailsOpen: boolean;
 }
 
 export const CharacterList = ({ characters, isDetailsOpen }: CharacterListProps) => {
-  const location = useLocation();
-  const navigate = useNavigate();
+  const { closeDetails } = useHandleDetails();
 
-  const closeDetails = () => {
-    const params = new URLSearchParams(location.search);
-    params.delete('details');
-    navigate(`/?${params.toString()}`);
-  };
-
-  return (
+  return characters.length > 0 ? (
     <ul className={classnames(styles.mainContainer, { [styles.small]: isDetailsOpen })} onClick={closeDetails}>
       {characters.map(character => (
         <CharacterItem key={character.name} character={character} isDetailsOpen={isDetailsOpen} />
       ))}
     </ul>
+  ) : (
+    <div className={styles.emptySearch}>Sorry, we couldn`t find anything matching your search.</div>
   );
 };
