+++ src/pages/home/home.module.scss
@@ -64,16 +64,6 @@
   padding: 0;
 }
 
-.empty-search {
-  @include flex(center, center);
-  @include font(2.8rem, 700, 1, 0.05rem);
-  width: 70%;
-  margin: 0 auto;
-  color: var(--color-white-default);
-  text-align: center;
-  flex-grow: 1;
-}
-
 .leftContainer {
   @include flex(flex-start, center, column);
 }
