--- src/store/middlewares/favoritesMiddleware.ts
@@ -1,7 +1,0 @@
-import { Middleware } from '@reduxjs/toolkit';
-
-export const favoritesNotifyMiddleware: Middleware = () => next => action => {
-  const result = next(action);
-
-  return result;
-};
