+++ src/tests/Search/Search.test.tsx
@@ -1,5 +1,5 @@
 import { Search } from '@components/Search/Search';
-import { useLocalStorage } from '@hooks/useSearchQuery';
+import { useLocalStorage } from '@hooks/useLocalStorage';
 import '@testing-library/jest-dom';
 import { render, screen } from '@testing-library/react';
 import { userEvent } from '@testing-library/user-event';
